+++ test/test_flagging.py
@@ -1,4 +1,5 @@
 import os
+import pytest
 import tempfile
 from unittest.mock import MagicMock
 
@@ -117,3 +118,19 @@ class TestDisableFlagging:
                 self.fail("launch() raised a PermissionError unexpectedly")
 
         io.close()
+
+
+class TestInterfaceConstructsFlagMethod:
+    @pytest.mark.parametrize(
+        "allow_flagging, called",
+        [
+            ("manual", True),
+            ("auto", True),
+            ("never", False),
+        ],
+    )    
+    def test_flag_method_init_called(self, allow_flagging, called):
+        flagging.FlagMethod.__init__ = MagicMock()
+        flagging.FlagMethod.__init__.return_value = None
+        gr.Interface(lambda x:x, "text", "text", allow_flagging=allow_flagging)
+        assert (flagging.FlagMethod.__init__.called == called)
